// src/services/dashboard/type.ts
export interface IBaseName {
  code: string;
  name: string;
}
// ...
export interface IGpaScore {
  id: number;
  educationYear: {
    code: string;
    name: string;
    current: boolean;
  };
  level: IBaseName;
  gpa: string;
  credit_sum: string;
  subjects: number;
  debt_subjects: number;
  method: 'one_year' | 'all_year';
}